+++ src/patch-repl.ts
@@ -1,27 +1,11 @@
 import repl, { type REPLServer, type REPLEval } from 'node:repl';
-import { transform } from 'esbuild';
+import backend from './backend';
 
 const patchEval = (nodeRepl: REPLServer) => {
 	const { eval: defaultEval } = nodeRepl;
 	const preEval: REPLEval = async function (code, context, filename, callback) {
 		try {
-			const transformed = await transform(
-				code,
-				{
-					sourcefile: filename,
-					loader: 'ts',
-					tsconfigRaw: {
-						compilerOptions: {
-							preserveValueImports: true,
-						},
-					},
-					define: {
-						require: 'global.require',
-					},
-				},
-			);
-
-			code = transformed.code;
+			code = await backend.replTransform(code, filename);
 		} catch {}
 
 		return defaultEval.call(this, code, context, filename, callback);
